requirements.txt
aiogram==3.17.0
apscheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
reportlab==4.2.0
pytz==2024.1